+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.12.5
 - CSS nicer layout
 - Prop not null switch
+- fix broken footer links
 
 # 0.12.4
 - SelectRef: Display selected properties to show instead of id (i.e. UUID)